Changed distance unit from "pixels per frame" to "pixels per second" Added explosion component Modified interfaces to have its methods optional Fixed bug: level updating time during pause Changed time unit in level from ms to s

+++ src/com/mitch/flyship/Enemy/Components/VerticalEnemy.java
@@ -41,7 +41,7 @@ public class VerticalEnemy extends EnemyComponent {
     }
 
     @Override
-    public void onUpdate(float deltaTime) {
+    public void onUpdate(double deltaTime) {
         super.onUpdate(deltaTime);
 
         Graphics g = enemy.getLevel().getAirshipGame().getGraphics();
