+++ src/com/mitch/flyship/objects/Button.java
@@ -66,7 +66,7 @@ public class Button extends GameBody {
 		currentImage = Assets.getImage(image);
 	}
 	@Override
-	public void onUpdate(float deltaTime) 
+	public void onUpdate(double deltaTime)
 	{
 		boolean touched = isTouched(offset);
 		boolean screenTouched = game.getInput().isTouchDown(0);
