+++ src/com/mitch/flyship/objects/Coin.java
@@ -38,7 +38,7 @@ public class Coin extends GameBody {
 	}
 	
 	@Override
-	public void onUpdate(float deltaTime) {
+	public void onUpdate(double deltaTime) {
 		
 		if (getPos().y > game.getGraphics().getHeight()) {
 			level.getBodyManager().removeBody(this);
